% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adw.R
\name{adw}
\alias{adw}
\title{adw function, Angular distance weighting interpolation

Gridding the irregularly-spaced data onto a regular latitude-longitude grid
by weighting each station according to its distance and angle from the centre
of a search radius}
\usage{
adw(
  dd,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  gridSize = 1,
  cdd = 1e+06,
  m = 4
)
}
\arguments{
\item{dd}{is the dataframe}

\item{xmin}{is the minimum of the longitude}

\item{xmax}{is the minimum of the longitude}

\item{ymin}{is the minimum of the latitude}

\item{ymax}{is the maximum of the latitude}

\item{gridSize}{is the grid resolution}

\item{cdd}{is the correlation decay distance, unit: meter}

\item{m}{is used to adjust the weighting function further}
}
\value{
a dataframe
}
\description{
adw function, Angular distance weighting interpolation

Gridding the irregularly-spaced data onto a regular latitude-longitude grid
by weighting each station according to its distance and angle from the centre
of a search radius
}
