% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDW.R
\name{IDW}
\alias{IDW}
\title{Inverse Distance Weighting interpolation}
\usage{
IDW(xy, z, xrange, yrange, res = 0.5, n.station = 5, p = 1)
}
\arguments{
\item{xy}{A matrix (N,2) with longitude and latitude of points of data observed}

\item{z}{A vector (N) with the values observeds in the points}

\item{xrange}{A vector with extremes longitudes of grid eg: c(-45,-35)}

\item{yrange}{A vector with extremes latitudes of grid. eg: c(-1,-5)}

\item{res}{The resolution of grid in degree}

\item{n.station}{Number of stations used per point for interpolation. The default is 8.}

\item{p}{Power parameter. The default is 1.}
}
\value{
Interpoled regular grid in Netcdf, data.frame or raster by ADW method.
}
\description{
Apply the Inverse Distance Weighting interpolation
method on the irregular points of meteorological observations to a regular grid.
}
\author{
Rodrigo Lins R Júnior
}
