% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetC.R
\name{NetC}
\alias{NetC}
\title{Convert data.frame for Netcdf}
\usage{
NetC(data, varName = "data", fileName = "V1", mask = F,
  path = NULL)
}
\arguments{
\item{data}{The data frame output by IDW/EDW/ADW function}

\item{varName}{Variable name in netcdf file}

\item{fileName}{Netcdf file name archive (or path with name). Default is "data" and the archive will be save in
current directory}

\item{mask}{A logical param for choose make a geografical mask by a shapefile. If TRUE, is necessary put 
a path with shapefile andress in any directory.}

\item{path}{The directory of shapefile for geografical mask in grid (Only mask=TRUE)}
}
\value{
Generate a netcdf file in current directory or in path directory
}
\description{
Convert the data frame from IDW/EDW/ADW function for a netcdf file
}
\author{
Rodrigo Lins R Júnior
}
