% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDW.R
\name{RDW}
\alias{RDW}
\title{Radial Distance Weighting interpolation}
\usage{
RDW(xy, z, xrange, yrange, res = 0.5, n.station = 5, p = 1, r = 10)
}
\arguments{
\item{xy}{A matrix (N,2) with longitude and latitude of points of data observed}

\item{z}{A vector (N) with the values observeds in the points}

\item{xrange}{A vector with extremes longitudes of grid eg: c(-45,-35)}

\item{yrange}{A vector with extremes latitudes of grid. eg: c(-1,-5)}

\item{res}{The resolution of grid in degree}

\item{n.station}{Number of stations used per point for interpolation. The default is 8.}

\item{p}{Power parameter. The default is 1.}

\item{r}{radius (in degree) for interpolation. The default ins 10}
}
\value{
A data.frame with longitude, latitude and interpoled points
}
\description{
Apply the Radial Distance Weighting interpolation
method on the irregular points of meteorological observations to a regular grid.
}
\examples{
data(TempBrazil) # Temperature for some poins of Brazil

LonLat=TempBrazil[,1:2] #Data.frame with Longtude and Latitude
Temp=TempBrazil[,3] # Vector with observations in points

LonInterval=c(-78,-34.10)  # Coordinates of extremes poins of longitude to grid
LatInterval=c(-36,5)  # Coordinates of extremes poins of latitude to grid

Interpoled=RDW(xy=LonLat,z=Temp,xrange = LonInterval,yrange = LatInterval)

}
\author{
Rodrigo Lins da Rocha Junior
}
