% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tavg}
\alias{tavg}
\title{Henan Province temperature data}
\format{
A data frame with 116 rows and 3 variables:
\describe{
\item{lon}{numeric, longitude}
\item{lat}{numeric, latitude}
\item{value}{numeric, average temperature value}
...
}
}
\usage{
tavg
}
\description{
Henan Province temperature data
}
\keyword{datasets}
