% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADW.R
\name{ADW}
\alias{ADW}
\title{Angular Distance Weighting interpolation}
\usage{
ADW(xy, z, xrange, yrange, res = 0.5, n.station = 8, m = 4,
  CDD = 50)
}
\arguments{
\item{xy}{A matrix (N,2) with longitude and latitude of points of data observed}

\item{z}{A vector (N) with the values observeds in the points}

\item{xrange}{A vector with extremes longitudes of grid eg: c(-45,-35)}

\item{yrange}{A vector with extremes latitudes of grid. eg: c(-1,-5)}

\item{res}{The resolution of grid in degree}

\item{n.station}{Number of stations used per point for interpolation. The default is 8.}

\item{m}{The param m of distance weight. The default is 4.}

\item{CDD}{Correlation distance decay param (Km).The default is 100 km.}
}
\value{
A data.frame with longitude, latitude and interpoled points
}
\description{
Apply the Angular Distance Weighting interpolation
method on the irregular points of meteorological observations to a regular grid.
}
\examples{
data(TempBrazil) # Temperature for some poins of Brazil


LonLat=TempBrazil[,1:2] #Data.frame with Longtude and Latitude
Temp=TempBrazil[,3] # Vector with observations in points

LonInterval=c(-78,-34.10)  # Coordinates of extremes poins of longitude to grid
LatInterval=c(-36,5)  # Coordinates of extremes poins of latitude to grid

Interpoled=ADW(xy=LonLat,z=Temp,xrange = LonInterval,yrange = LatInterval)

}
\author{
Rodrigo Lins R. Jr., Fabricio  Daniel S. S.
}
